+++ src/app/App.js
@@ -4,7 +4,7 @@ import { Route, Switch, Redirect } from "react-router-dom";
 import Products from "./layouts/products";
 import Login from "./layouts/login";
 import Main from "./layouts/main";
-import NavBar from "./components/navBar";
+import NavBar from "./components/ui/navBar";
 
 const App = () => {
     return (
