+++ src/app/components/page/productPage/index.js
@@ -1,0 +1,2 @@
+import ProductPage from "./productPage";
+export default ProductPage;
