+++ src/app/components/page/productsListPage/index.js
@@ -1,0 +1,2 @@
+import ProductsListPage from "./productsListPage";
+export default ProductsListPage;
