+++ src/app/components/page/productsListPage/productsListPage.jsx
@@ -1,16 +1,16 @@
 import React, { useState, useEffect } from "react";
 import PropTypes from "prop-types";
 
-import { paginate } from "../utils/paginate";
-import Pagination from "./pagination";
-import api from "../api";
-import GroupList from "./groupList";
-import SearchStatus from "./searchStatus";
-import ProductsTable from "./productsTable";
+import { paginate } from "../../../utils/paginate";
+import Pagination from "../../common/pagination";
+import api from "../../../api";
+import GroupList from "../../common/groupList";
+import SearchStatus from "../../ui/searchStatus";
+import ProductsTable from "../../ui/productsTable";
 
 import _ from "lodash";
 
-const ProductsList = () => {
+const ProductsListPage = () => {
     const [currentPage, setCurrentPage] = useState(1);
     const [categories, setCategories] = useState();
     const [searchQuery, setSearchQuery] = useState("");
@@ -143,11 +143,11 @@ const ProductsList = () => {
             </div>
         );
     }
-    return "Loading productsList.jsx";
+    return "Loading productsListPage.jsx";
 };
 
-ProductsList.propTypes = {
+ProductsListPage.propTypes = {
     products: PropTypes.array
 };
 
-export default ProductsList;
+export default ProductsListPage;
