+++ src/app/layouts/login.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import LoginForm from "../components/loginForm";
+import LoginForm from "../components/ui/loginForm";
 
 const Login = () => {
     return <LoginForm />;
