+++ src/app/layouts/products.jsx
@@ -1,12 +1,14 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import ProductPage from "../components/productPage";
-import ProductsList from "../components/productsList";
+import ProductPage from "../components/page/productPage";
+import ProductsListPage from "../components/page/productsListPage";
 
 const Products = () => {
     const params = useParams();
     const { prodId } = params;
-    return <>{prodId ? <ProductPage prodId={prodId} /> : <ProductsList />}</>;
+    return (
+        <>{prodId ? <ProductPage prodId={prodId} /> : <ProductsListPage />}</>
+    );
 };
 
 export default Products;
